vpdupdate: Find the PCI/USB ids files at runtime
Currently the pci.ids and usb.ids files are located by the Makefile
at compile time. This presents a problem when host machine has a
different location for the files. This patch addresses this problem
by checking for the above files in the possible paths at runtime.

Changes since V1:

 * Remove the obsolete ids look up at build time.
 * Set a non-empty value for the prefix string if we are unable
   to find the ids files, to prevent the repeated lookups.
 * Fixed comment styles.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>
Signed-off-by: Phani Yadav <phayadav@in.ibm.com>
Signed-off-by: Suzuki K. Poulose <suzuki@in.ibm.com>