lsvpd: fix path issue with -z option
Fixing the path issue in lsvpd -z option when using
a db file to retrive the vpd information. Also
removing the message "Run vpdupdate" when we try to
retrieve the vpd  information from a db archive file.
Providing  "Run vpdupdate" message to user is not
relevant in this context as we have already mentioned the file.

Reported-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Phani Yadav <phayadav@linux.vnet.ibm.com>
Acked-by: Suzuki K Poulose <suzuki@in.ibm.com>