lsvpd: RtasCollector extends PlatformCollector
This patch is to add inheritance to PlatformCollector for RtasCollector class
and also encapsulate common functions. The reason being, Rtas is not longer
supported on PowerKVM Host. Hence, vpdupdate tool which makes a call to
RtasCollector is no longer relevant.

Thus PlatformCollector parent class has been introduced which will return
instance of OpalCollector on PowerNV platform or RtasPlatformCollector on
pSeries platform.

Signed-off-by: Bharani C.V <bharanve@linux.vnet.ibm.com>
Acked-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>