Sanitize VPD Field before updating in the DB
This patch checks if value being updated in db
is in binary form or string. If it is binary,
then hexdump is stored as string and updated
in vpd.db. This is done to avoid displaying junk values
for fields with binary data.

Some fields may have trailing '0's which is handled.
Some fields have full of space characters. Convert them
to hexa.

Signed-off-by: Phani Yadav <phayadav@linux.vnet.ibm.com>
Signed-off-by: Bharani C.V <bharanve@linux.vnet.ibm.com>
Signed-off-by: Suzuki K. Poulose <suzuki@in.ibm.com>