vpdupdate: Threshold VPD_CHANGED return codes
The rtas_get_vpd() could return a VPD_CHANGED code to the caller
to indicate that there are some changes in the VPD data and hence
we need to restart the call from sequence 1. However, on a buggy
firmware we may get this problem and we keep looping forever.

To avoid this, we set a threshold count for the number of consecutive
VPD_CHANGED to decide to quit.

Signed-off-by: Suzuki K. Poulose <suzuki@in.ibm.com>