lsmcode: fix path issue with -z option
Fixing the path issue in lsmcode -z option when using a db file
to retrieve the vpd information. Also removing the message "Run
vpdupdate" when we try to retrieve the vpd  information from a
db archive file. Providing  "Run vpdupdate" message to user is
not relevant in this context as we have already mentioned the file.

Signed-off-by: Janani Venkataraman <jananive@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>