lsvpd: Defer the freeing of devSg
After calling basename, we need to defer freeing the pointer.

This is because the basename function may return pointers to statically
allocated memory which may be overwritten by subsequent calls. They may
return a pointer to some part of path and hence the string refered to
by the path  should not be modified or freed until the pointer returned
by the function is no longer required.

Signed-off-by: Janani Venkataraman <jananive@linux.vnet.ibm.com>
Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>
Reviewed-by: Naveen N. Rao <naveen.n.rao@linux.vnet.ibm.com>
[Removed redundant return - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>