getdircontent function is modified for error checking condition
getDirContents function takes input as a string and convert it to char * and
then copies it to a static buffer.
We can replace static array to a string object, which will fix the issue of
static memory size check, as string size will get increased or decreased
dynamically.
fs_isFile is an overloaded function with parameter string and char*, so we can
call that function by passing string as a parameter.

Signed-off-by: Ankit Kumar <ankit@linux.vnet.ibm.com>
Reviewed-by: Vaibhav Jain <vaibhav@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>