lsvpd: Fix uninitialized variable use
fs_isDir, fs_isLink, fs_isFile functions takes path as a param and checks
whether given path is a dir, link or file. It returns true if path provided
is a dir, link, file respectively else it returns false.

e.g: In function fs_isFIle:
If path exists but it is not a file(may be it is a dir or link), in that case
this function returns garbage value as ret is not initialized.

This patch fixes above issue by initializing the return variable with false.

This function match:
value ret is initialized to string.length() but is never used.
This patch removed var. initialization as it is latter assigned before use.

Signed-off-by: Ankit Kumar <ankit@linux.vnet.ibm.com>
Reviewed-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>