lsvpd: Unique name for temp Device file given and file closed before deleting it.
This piece of code used mkstemp to get unique name of file in that particular
directory. As file name is not hardcoded, so parameter for device_close and
device_open is changed to pass device location path.

This patch also fixes the issue of opened file which is not closed by closing
file descriptor.

Signed-off-by: Ankit Kumar <ankit@linux.vnet.ibm.com>
[Removed redundant variable - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>