lsvpd: Fixes Issue of var assigned but not used and unused variable
This patch fixes the issue of variable which is assigned but not used or
assigned again before use and removed unused variable.

In function match in sysfs_SCSI_Fill.cpp :
string::npos is of type size_t and string.find() also returns size_t, so we
don't need to typecast the returned value and stored first and then compare.
In Function fillSCSIComponent:
We don't use the return value of function get_mm_scsi, so no need to store the
return value in any variable.

Signed-off-by: Ankit Kumar <ankit@linux.vnet.ibm.com>
[Removed remaining unused variable in devicetreecollector.cpp - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>