lsvpd: Faulty logical expression corrected
Mentioned call will lookup for subdevice type and finds which is actual
matching device and if it doesn't find devices it will have key value which
will have value as -1 and it shows that device didn't match.
To verify above logic, present code is a buggy code which always gives a
result that device found.
This patch fixes above mentioned issue in error condition.

Signed-off-by: Ankit Kumar <ankit@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>