Make device_open/close function part of lsvpd namespace
device_open and device_close functions are defined in sysfs_SCSI_Fill.cpp,
which has scsi related functionality. It is not declared in any header
file, hence will have limited static scope inside the file where it is defined
and cann't be used in other functions outside the file.

Since it is a generic function used to open and close device file. So it should
be used from any part of this package. This patch fixes above mentioned issue by
making device_open/close functions part of lsvpd namespace.

Signed-off-by: Ankit Kumar <ankit@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>