lsmcode: Fix failure of displaying all device possible micro code
If OPAL firmware failed to provide below product version string printSystem
currently returns false in case of all(--All). Due to which printVPD just
returns instead of continue to provide microcode levels of other leaf devices.

open-power       "SUPERMICRO-P9DSU-V1.03-20180205-imp"

This patch fixes this issue by not checking the return code of printSystem
in case of all option, so it displays microcode levels of as much devices
as possible(--all).

Signed-off-by: Pridhiviraj Paidipeddi <ppaidipe@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>