Enable NVME f1h log page VPD information parsing
The patch implements parsing for NVME devices
with f1h log page capability.

The log page is optional and part of the nvme device file,
can be accessed using ioctl.

The patch introduces the nvme log page configuration file
which can be used to fetch the format of the f1h page vpd
data based on the log page version.

The configuration file format is

"logpageVersion", "_:4,fieldKeyword1:keywordLength1,fieldKeyword2:keywordLength2..."

The first field in vpd logpage is always log page version
of 4 character length, so format starts with "_:4".

For example:
For logpage version "0001" the fields "Description(MN)" is
of size 40 characters followed by "Master PN" of size 12
characters and so on. So the config file is defined as
"0001", "_:4,MN:40,PN:12..."

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>
[Added few return value validation, removed log message if we
 fail to read f1 page - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>