lsvpd: Add Recommends clause to lsvpd configuration file
The solution to add a "Recommends" clause was recommended by distro
as part of solution for a customer reported issue where usb.ids file
was not found by lsvpd command, but that is available from the hwdata
package.

This patch adds the hwdata package as a recommended package for lsvpd.

Signed-off-by: Brahadambal Srinivasan <latha@linux.vnet.ibm.com>
[We always assumed hwdata was installed by default. Without hwdata we
 will not be able to get various PCI/USB device details. We have
 hardcoded those ids files path. Hence build process will not be able
 to detect the dependency. - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>