devicetreecollector: buildSCSI() - add support for SCSI loc-code
fillQuickVPD failing to find the loc-code, doesn't always attribute to
the missing device-tree entry associated with the sysfs device path.
There are also configurations like USB hubs, to which the SCSI devices
gets connected with special device-tree layout. A USB controller, that
supports both USB 3.0 and USB 2.0 devices will create two USB hubs for
communicating with devices. Where one hub controllers devices with
speeds ranging upto  480M (USB 2.0) and another for devices operating
at the  5000M (USB 3.0). The layout might be little complex when the
USB 3.0 devices are internal on board ones and 2.0 is an external
device.

                     root hub
                    /         \
           USB hub 1 (2.0)  USB hub 2 (3.0)
                   |         /            \
                 external   internal    internal
                  device    device1     device2

the populated device-tree adds both of the internal devices
under the hub1, that hosts the slower external device, with
the layout:
usb@1
|_ hub@1
   |_ cdrom@1 (USB 2.0)
   |_ usb-scsi@6 (USB 3.0, internal device 1)
   |_ usb-scsi@8 (USB 3.0, internal device 2)

this patch add support to find location code for such SCSI devices
connected to USB hubs in a mix and match of external and internal
devices.

Signed-off-by: Kamalesh Babulal <kamalesh@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>