lsvpd: Fill firmware information by recursively looking for firmware specific files within sysfs ID
Currently, fillFirmware() is filling firmware information based on the
classNode value of a given component. It then looks for firmware specific
sysfs files (fw_version, fwrev, firmware_rev), within the classNode path,
to set firmware fields accordingly. For components that do not have a
classNode value, fillFirmware() does not look for firmware information
within the sysfs firmware files, and hence does not set firmware fields.

This patch recursively looks for sysfs firmware files within the ID path of
a component, and sets the firmware fields accordingly, if not already set.

Reported-by: Borislav Stoymirski <borislav.stoymirski@bg.ibm.com>
Signed-off-by: Sathvika Vasireddy <sv@linux.ibm.com>
Signed-off-by: Mahesh Salgaonkar <mahesh@linux.ibm.com>
Tested-by: Borislav Stoymirski <borislav.stoymirski@bg.ibm.com>