lsvpd/travis-ci: Fix Fedora 37 build dependency error
The Fedora 37 CI build fails with the following error:
error: Failed build dependencies:
  systemd-rpm-macros is needed by lsvpd-1.7.15-1.fc37.ppc64le

This occurs because the BuildRequires for systemd-rpm-macros in the
spec file cannot be satisfied as the package is not available in the CI
build environment.

Fix by explicitly installing systemd-rpm-macros in the Dockerfile
to ensure the dependency is available during the RPM build process.

Signed-off-by: Sathvika Vasireddy <sv@linux.ibm.com>